FPAD-8430: Separate txma handler lambda and logic into dedicated files
